Design ticketmetas

+++ app/src/main/java/edu/ucne/composedemo/data/remote/dto/TicketMetaResponseDto.kt
@@ -2,7 +2,7 @@ package edu.ucne.composedemo.data.remote.dto
 
 data class TicketMetaResponseDto(
     val id: Int,
-    val idTicket: Double?,
+    val idTicket: Int?,
     val asunto: String?,
     val empresa: String?,
     val estatus: Int?,
